+++ WHATS-NEW.md
@@ -1,2 +1,10 @@
-v.1.2
+v1.2
     - New visual improvements, added a "Hide" button on the scoreboard, and also enabled a public scoreaboard.
+
+v1.3
+    - Profile statistics are now viewable, minor frontend changes.
+    - Designed error pages
+    - User-agent detection instead of resolution for phones
+    - Badges for users to earn
+    - Identicon and profile pictures for users
+